+++ src/short_con/__init__.py
@@ -1,3 +1,10 @@
-from .main import cons, enumcons, constants, ShortConError
 from .version import __version__
+
+from .main import (
+    ShortConError,
+    cons,
+    constants,
+    dc,
+    enumcons,
+)